separate motor_model.py
