+++ .github/workflows/calibrate-motor.yml
@@ -8,6 +8,7 @@ on:
       - data/motortestdata.zip
       - calibrate/calibrate.py
       - calibrate/filter.py
+      - calibrate/motor_model.py
       - calibrate/calibrate_cli.py
       - pyproject.toml
 
@@ -32,6 +33,7 @@ jobs:
             calibrate/calibrate.py
             calibrate/calibrate_cli.py
             calibrate/filter.py
+            calibrate/motor_model.py
             pyproject.toml
             data/motortestdata.zip
           sparse-checkout-cone-mode: false
