+++ calibrate/calibrate_cli.py
@@ -1,6 +1,6 @@
 import argparse
 import glob
-from os.path import basename
+from os.path import basename, splitext
 from calibrate import *
 
 
@@ -24,7 +24,7 @@ def parse_args(src=None):
 
 def load_runs(filenames):
     testdata = [loadRun(f) for f in sorted([g for f in filenames for g in glob.glob(f)])]  # :P
-    print("Files:", [basename(data["filename"]) for data in testdata])
+    print("Files:", ', '.join([splitext(basename(data["filename"]))[0] for data in testdata]))
     return testdata
